Minor notes. Added README.
I established this branch to do "time shifting," i.e., but it's already done. The Garmin dates the files according to the local time set in the GPS which is what I want.

Forgot to delete the notes about that in the main file. Will do it after merging this branch back in. Don't want to get confused by recommiting.